# Problem : getting id_asset_element is not optimized