@geraldguillaume 2018-06-12
Last comment : it is also aligned with that code (redundant ??) https://github.com/42ity/fty-rest/blob/master/src/db/assets/assetr.cc#L75